dataflow/cloud-functions-daily-data/requirements.txt
yfinance
datetime
google-api-python-client
google-cloud-storage